+++ activerecord/lib/active_record/persistence.rb
@@ -7,6 +7,10 @@ module ActiveRecord
   module Persistence
     extend ActiveSupport::Concern
 
+    included do
+      extend ActiveModel::AttributeAssignment::ClassMethods
+    end
+
     module ClassMethods
       # Creates an object (or multiple objects) and saves it to the database, if validations pass.
       # The resulting object is returned whether the object was saved successfully to the database or not.
@@ -54,36 +58,6 @@ module ActiveRecord
           object = new(attributes, &block)
           object.save!
           object
-        end
-      end
-
-      # Builds an object (or multiple objects) and returns either the built object or a list of built
-      # objects.
-      #
-      # The +attributes+ parameter can be either a Hash or an Array of Hashes. These Hashes describe the
-      # attributes on the objects that are to be built.
-      #
-      # ==== Examples
-      #   # Build a single new object
-      #   User.build(first_name: 'Jamie')
-      #
-      #   # Build an Array of new objects
-      #   User.build([{ first_name: 'Jamie' }, { first_name: 'Jeremy' }])
-      #
-      #   # Build a single object and pass it into a block to set other attributes.
-      #   User.build(first_name: 'Jamie') do |u|
-      #     u.is_admin = false
-      #   end
-      #
-      #   # Building an Array of new objects using a block, where the block is executed for each object:
-      #   User.build([{ first_name: 'Jamie' }, { first_name: 'Jeremy' }]) do |u|
-      #     u.is_admin = false
-      #   end
-      def build(attributes = nil, &block)
-        if attributes.is_a?(Array)
-          attributes.collect { |attr| build(attr, &block) }
-        else
-          new(attributes, &block)
         end
       end